requirements.txt
ipython==7.34.0
matplotlib==3.5.3
numpy==1.21.6
Pillow==9.5.0
scikit_learn==1.0.2
scipy==1.7.3
torchvision~=0.11.0
tqdm==4.65.0
